Winston Day 19 Minor Removed State Tree and Added Player Functions
State tree is very unstable and still experimental, products should not be shipped. state trees also cannot interact with ai controller, thus no access to EQS thus the entity is blind and deaf to world stimuli.
Replaced it with FSM and Enum based behaviors. simplifying the AI further to improve performance.
AI to store data in the body and update the controller in order to reduce the number of blueprints used in the content drawer.